Break out individual & common Python-unit-test-framework unit tests.
- improve readability of our unit tests
- improve intent of our unit tests
- remove use of `.skip()` to suite test execution logic
- remove unnecessary tslint warning suppressions
